pyproject.toml
[project]
name = "docu-chan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
